backend/requirements-dev.txt
pytest
pytest-django
pytest-asyncio
pytest-mock
coverage
requests-mock
freezegun
flake8
requests
mock
